requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
sqlite3
unidecode
dateparser
numpy